requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.1
opencv-python==4.8.1.78
scipy==1.11.3
rembg==2.0.50
python-dateutil==2.8.2
zipfile36==0.1.3
imagehash==4.3.1